refactor: move response models to Application layer for clean architecture
