+++ c#/time-deposit-kata-net/Application/ITimeDepositService.cs
@@ -6,6 +6,6 @@ namespace time_deposit_kata_net.Application
     public interface ITimeDepositService
     {
         Task UpdateAllBalancesAsync();
-        Task<List<TimeDeposit>> GetAllTimeDepositsAsync();
+        Task<List<TimeDepositResponse>> GetAllTimeDepositsAsync();
     }
 }
